Fixed Dynare signature.

+++ build.sh
@@ -31,8 +31,9 @@ set -ex
 THISSCRIPT=$(readlink -f $0)
 ROOT_DIRECTORY=`dirname $THISSCRIPT`
 
-# Set options.
+# Set defaults for options.
 FORCE_BUILD=0
+SIGN_DYNARE=0
 
 # Set the number of threads
 NTHREADS=`nproc --all`
@@ -221,9 +222,13 @@ if [ $BUILD_WINDOWS_EXE -eq 1 ]; then
     # Create Windows installer
     cd $THIS_BUILD_DIRECTORY/windows
     makensis dynare.nsi
-    $ROOT_DIRECTORY/signature/osslsigncode sign -pkcs12 $ROOT_DIRECTORY/dynare-object-signing.p12 -n Dynare -i http://www.dynare.org -in dynare-$VERSION-win.exe -out dynare-$VERSION-win-signed.exe
-    rm dynare-$VERSION-win.exe
-    mv dynare-$VERSION-win-signed.exe $ROOT_DIRECTORY/win/dynare-$VERSION-win.exe
+    if [ $SIGN_DYNARE -eq 1 -a ! -f "$ROOT_DIRECTORY/impossible-to-sign-dynare" ]; then
+        $ROOT_DIRECTORY/signature/osslsigncode sign -pkcs12 $ROOT_DIRECTORY/dynare-object-signing.p12 -n Dynare -i http://www.dynare.org -in dynare-$VERSION-win.exe -out dynare-$VERSION-win-signed.exe
+        rm dynare-$VERSION-win.exe
+        mv dynare-$VERSION-win-signed.exe $ROOT_DIRECTORY/win/dynare-$VERSION-win.exe
+    else
+	mv dynare-$VERSION-win.exe $ROOT_DIRECTORY/win/dynare-$VERSION-win.exe
+    fi
     ln --relative --symbolic --force $ROOT_DIRECTORY/win/dynare-$VERSION-win.exe $ROOT_DIRECTORY/win/dynare-latest-win.exe
     cd $THIS_BUILD_DIRECTORY
 fi
